refactor(F113): extract segment parsing + icons (R3 P2#4 part 2)
Extracted from DirectoryBrowser.tsx:
- directory-segments.ts: windowsDriveSegments, pathToSegments, buildBrowseUrl,
  shouldFallbackToHome + BrowseEntry/BrowseResult/BreadcrumbSegment types
- directory-browser-icons.tsx: HomeIcon, PcIcon, DriveIcon, FolderIcon,
  TerminalIcon

DirectoryBrowser.tsx: 583 -> 437 lines. Still above the 350-line hard limit;
breadcrumb/drives-view sub-components extraction follows. tsc clean, 17/17
tests green.
